Clean up module
