+++ distance3d/pressure_field/__init__.py
@@ -1,174 +1,21 @@
 """Pressure field model for contact wrenches."""
-import aabbtree
-import numpy as np
-from ..utils import invert_transform
-from ..benchmark import Timer
-from ._tetra_mesh_creation import make_tetrahedral_cube, make_tetrahedral_icosphere
-from ._mesh_processing import tetrahedral_mesh_aabbs, center_of_mass_tetrahedral_mesh
+from ._interface import contact_forces, find_contact_surface
 from ._rigid_body import RigidBody
 from ._contact_surface import ContactSurface
-from ._tetrahedron_intersection import intersect_tetrahedron_pairs, intersect_tetrahedron_pair, intersect_halfplanes
-from ._forces import contact_surface_forces, accumulate_wrenches, compute_contact_force
-from ._halfplanes import plot_halfplanes_and_intersections
+
+from ._halfplanes import intersect_halfplanes, plot_halfplanes_and_intersections
+from ._tetra_mesh_creation import make_tetrahedral_icosphere
+from ._mesh_processing import center_of_mass_tetrahedral_mesh, tetrahedral_mesh_aabbs
+from ._barycentric_transform import barycentric_transforms
+from ._tetrahedron_intersection import intersect_tetrahedron_pair
+from ._forces import compute_contact_force
 
 
-def contact_forces(rigid_body1, rigid_body2, return_details=False, timer=None):
-    """Contact forces between two objects.
-
-    Parameters
-    ----------
-    rigid_body1 : RigidBody
-        First rigid body.
-
-    rigid_body2 : RigidBody
-        Second rigid body.
-
-    return_details : bool
-        Return additional contact details.
-
-    timer : Timer, optional (default: None)
-        Measures time to execute certain parts of this function.
-
-    Returns
-    -------
-    intersection : bool
-        Do both bodies intersect?
-
-    wrench12_in_world : array, shape (6,)
-        Forces and torques caused by body 1 acting on body 2 in world frame.
-
-    wrench21_in_world : array, shape (6,)
-        Forces and torques caused by body 2 acting on body 1 in world frame.
-
-    details : dict, optional
-        Additional contact details.
-    """
-    if timer is None:
-        timer = Timer()
-
-    contact_surface = find_contact_surface(rigid_body1, rigid_body2, timer)
-
-    timer.start("accumulate_wrenches")
-    wrench12_in_world, wrench21_in_world = accumulate_wrenches(
-        contact_surface, rigid_body1, rigid_body2)
-    timer.stop_and_add_to_total("accumulate_wrenches")
-
-    if return_details:
-        timer.start("make_details")
-        details = contact_surface.make_details(
-            rigid_body1.tetrahedra_points, rigid_body2.tetrahedra_points)
-        timer.stop_and_add_to_total("make_details")
-        return contact_surface.intersection, wrench12_in_world, wrench21_in_world, details
-    else:
-        return contact_surface.intersection, wrench12_in_world, wrench21_in_world
-
-
-def find_contact_surface(rigid_body1, rigid_body2, timer=None):
-    """Find contact plane of two rigid bodies.
-
-    Note that this function will transform rigid_body1 into the frame of
-    rigid_body2.
-
-    Parameters
-    ----------
-    rigid_body1 : RigidBody
-        First rigid body.
-
-    rigid_body2 : RigidBody
-        Second rigid body.
-
-    timer : Timer, optional (default: None)
-        Measures time to execute certain parts of this function.
-
-    Returns
-    -------
-    contact_surface : ContactSurface
-        Contact information.
-    """
-    if timer is None:
-        timer = Timer()
-
-    timer.start("transformation")
-    # We transform vertices of rigid_body1 to rigid_body2 frame to be able to
-    # reuse the AABB tree of rigid_body2.
-    body12body2 = np.dot(invert_transform(rigid_body2.body2origin_),
-                         rigid_body1.body2origin_)
-    rigid_body1.transform(body12body2)
-    timer.stop_and_add_to_total("transformation")
-
-    timer.start("broad_phase_tetrahedra")
-    broad_tetrahedra1, broad_tetrahedra2, broad_pairs = broad_phase_tetrahedra(
-        rigid_body1, rigid_body2)
-    timer.stop_and_add_to_total("broad_phase_tetrahedra")
-
-    timer.start("barycentric_transform")
-    unique_indices1 = np.unique(broad_tetrahedra1)
-    unique_indices2 = np.unique(broad_tetrahedra2)
-    X1 = barycentric_transforms(rigid_body1.tetrahedra_points[unique_indices1])
-    X2 = barycentric_transforms(rigid_body2.tetrahedra_points[unique_indices2])
-    X1 = {j: X1[i] for i, j in enumerate(unique_indices1)}
-    X2 = {j: X2[i] for i, j in enumerate(unique_indices2)}
-    timer.stop_and_add_to_total("barycentric_transform")
-
-    timer.start("intersect_tetrahedron_pairs")
-    intersection_result = intersect_tetrahedron_pairs(
-        broad_pairs, rigid_body1.tetrahedra_points, rigid_body2.tetrahedra_points,
-        rigid_body1.tetrahedra_potentials, rigid_body2.tetrahedra_potentials,
-        X1, X2)
-    contact_surface = ContactSurface(rigid_body2.body2origin_, *intersection_result)
-    timer.stop_and_add_to_total("intersect_tetrahedron_pairs")
-
-    timer.start("contact_surface_forces")
-    contact_areas, contact_coms, contact_forces, contact_triangles = contact_surface_forces(
-        contact_surface, rigid_body1)
-    contact_surface.add_polygon_info(
-        contact_areas, contact_coms, contact_forces, contact_triangles)
-    timer.stop_and_add_to_total("contact_surface_forces")
-
-    return contact_surface
-
-
-def broad_phase_tetrahedra(rigid_body1, rigid_body2):
-    """Broad phase collision detection of tetrahedra."""
-
-    # TODO fix broad phase for cube vs. sphere
-    # TODO speed up broad phase
-    # TODO store result in RigidBody
-    """
-    aabbs1 = tetrahedral_mesh_aabbs(rigid_body1.tetrahedra_points)
-    aabbs2 = tetrahedral_mesh_aabbs(rigid_body2.tetrahedra_points)
-    tree2 = aabbtree.AABBTree()
-    for j, aabb in enumerate(aabbs2):
-        tree2.add(aabbtree.AABB(aabb), j)
-    broad_tetrahedra1 = []
-    broad_tetrahedra2 = []
-    for i, aabb in enumerate(aabbs1):
-        new_indices2 = tree2.overlap_values(aabbtree.AABB(aabb))
-        broad_tetrahedra2.extend(new_indices2)
-        broad_tetrahedra1.extend([i] * len(new_indices2))
-
-    broad_pairs = list(zip(broad_tetrahedra1, broad_tetrahedra2))
-    assert len(broad_tetrahedra1) == len(broad_tetrahedra2)
-    for i, aabb in enumerate(aabbs1):
-        if tree2.does_overlap(aabbtree.AABB(aabb)):
-            assert i in broad_tetrahedra1
-        else:
-            assert i not in broad_tetrahedra1
-
-    return broad_tetrahedra1, broad_tetrahedra2, broad_pairs
-    """
-
-    # FIXME workaround for broad phase bug:
-    from itertools import product
-    broad_tetrahedra1 = np.array(list(range(len(rigid_body1.tetrahedra_))), dtype=int)
-    broad_tetrahedra2 = np.array(list(range(len(rigid_body2.tetrahedra_))), dtype=int)
-    broad_pairs = list(product(broad_tetrahedra1, broad_tetrahedra2))
-    return broad_tetrahedra1, broad_tetrahedra2, broad_pairs
-
-
-def barycentric_transforms(tetrahedra_points):
-    """Returns X. X.dot(coords) = (r, 1), where r is a Cartesian vector."""
-    # NOTE that in the original paper it is not obvious that we have to take
-    # the inverse
-    return np.linalg.pinv(np.hstack((tetrahedra_points.transpose((0, 2, 1)),
-                                     np.ones((len(tetrahedra_points), 1, 4)))))
+__all__ = [
+    "contact_forces", "find_contact_surface", "RigidBody", "ContactSurface",
+    # exported for unit tests and specific examples:
+    "intersect_halfplanes", "plot_halfplanes_and_intersections",
+    "make_tetrahedral_icosphere", "center_of_mass_tetrahedral_mesh",
+    "tetrahedral_mesh_aabbs", "barycentric_transforms",
+    "intersect_tetrahedron_pair", "compute_contact_force"
+]
